Fix the GCC warnings coming from modular airport code
The Linux release build emitted 24 warnings attributable to this branch.
All five distinct sites are addressed without suppressing anything:

- TeleportAircraftOnModularTile was declared in both aircraft.h and
  modular_airport_cmd.h, and the latter includes the former, so every
  translation unit reaching it got -Wredundant-decls (21 occurrences).
  Dropped the aircraft.h copy: the function is defined in
  modular_airport_cmd.cpp and its declaration belongs beside the other
  modular entry points. The sole caller, modular_airport_build.cpp,
  already includes modular_airport_cmd.h.
- modular_airport_gui.cpp forward-declared CcBuildAirport, which
  airport_cmd.h (already included) declares as a CommandCallback.
- Two -Wsign-compare sites in AirportMoveModular: taxi_path_index is
  uint16_t, so `index + 1` promotes to int and was compared against
  vector::size_type. Widened to size_t; the comparison is unchanged for
  every representable value.
- airport_template_gui.cpp mixed the AirportTiles enumerator APT_APRON
  with a uint8_t in a conditional (-Wextra).

Unit tests pass (6183 assertions, 42 cases) and all three regression
fixtures clear their floors: mass7-inair 9251, helis2 13816, T5j2 6309.

+++ src/airport_template_gui.cpp
@@ -242,7 +242,7 @@ static bool BuildTemplateFromStation(const Station *st, AirportTemplate &templ)
  */
 static const DrawTileSprites *GetTileLayoutForTemplateTile(const AirportTemplateTile &t)
 {
-	uint8_t gfx = t.piece_type >= NEW_AIRPORTTILE_OFFSET ? APT_APRON : t.piece_type;
+	uint8_t gfx = t.piece_type >= NEW_AIRPORTTILE_OFFSET ? static_cast<uint8_t>(APT_APRON) : t.piece_type;
 	return GetAirportTileLayoutWithModularOverrides(gfx, t.piece_type, t.rotation, 0);
 }
